refactor: 기존 post 로직 제거
