--- src/application/post/type.ts
@@ -1,3 +1,0 @@
-import { HEADERS_OF_CONTENTS } from './constant'
-
-export type HeadersOfContents = typeof HEADERS_OF_CONTENTS[number]
