--- src/domain/owner/constant.ts
@@ -1,17 +1,0 @@
-export const KO_NAME = {
-  first: '하림',
-  last: '조',
-} as const
-
-export const EN_NAME = {
-  first: 'harim',
-  last: 'Cho',
-} as const
-
-export const BLOG = {
-  shortName: 'RIM',
-  fullName: 'RootInMeaningful',
-  domain: 'https://rootinmeaningful.vercel.app',
-} as const
-
-export const GITHUB_URL = 'https://github.com/Choharim'
